+++ lib/ooxml_parser/xlsx_parser/workbook/worksheet/table_part/extension_list/extension.rb
@@ -4,6 +4,7 @@ require_relative 'extension/data_validations'
 require_relative 'extension/sparkline_groups'
 require_relative 'extension/x14_table'
 require_relative 'extension/conditional_formattings'
+require_relative 'extension/x14_data_field'
 module OoxmlParser
   # Class for `ext` data
   class Extension < OOXMLDocumentObject
@@ -15,6 +16,8 @@ module OoxmlParser
     attr_accessor :table
     # @return [SparklineGroups] list of groups
     attr_reader :sparkline_groups
+    # @return [X14DataField] pivot data field in x14 namespace
+    attr_accessor :data_field
 
     # Parse Extension data
     # @param [Nokogiri::XML:Element] node with Extension data
@@ -30,6 +33,8 @@ module OoxmlParser
           @table = X14Table.new(parent: self).parse(column_node)
         when 'sparklineGroups'
           @sparkline_groups = SparklineGroups.new(parent: self).parse(column_node)
+        when 'dataField'
+          @data_field = X14DataField.new(parent: self).parse(column_node)
         end
       end
       self